new route ts file, use new file
