+++ src/components/Footer/Footer.tsx
@@ -2,7 +2,8 @@ import Link from "next/link";
 import ReactGA from "react-ga4";
 import { useQuery } from "react-query";
 import { getLatestDataReleaseInformation } from "../../apis/AggregatedData.api";
-import { Route, routes, routesWithGAEvents } from "../../utils/routes";
+import { type Route } from "../../types/Routes";
+import { routes, routesWithGAEvents } from "../../utils/routes";
 import ActiveLink from "../ActiveLink/ActiveLink";
 import SocialMediaIcons from "../Icons/SocialMediaIcons";
 import Logotype from "../Logotype/Logotype";
