+++ src/components/Navbar/Navbar-mobile/Navbar-mobile.tsx
@@ -3,8 +3,9 @@ import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 import ReactGA from "react-ga4";
 import { NavProps } from "../../../../types/globalTypes";
+import { type Route } from "../../../types/Routes";
 import handleBodyClass from "../../../utils/handleBodyClass";
-import { Route, routesWithGAEvents } from "../../../utils/routes";
+import { routesWithGAEvents } from "../../../utils/routes";
 import ActiveLink from "../../ActiveLink/ActiveLink";
 import CloseIcon from "../../Icons/CloseIcon/CloseIcon";
 import Logotype from "../../Logotype/Logotype";
