+++ src/types/Routes.ts
@@ -1,0 +1,11 @@
+export type Route = {
+	path?: string;
+	name: string;
+	secondary?: boolean;
+	opensNewTab?: boolean;
+	children?: {
+		path: string;
+		name: string;
+		opensNewTab?: boolean;
+	}[];
+};
